requirements.txt
openai-whisper==20250625
torch==2.1.0
numpy==1.26.4
pyaudio==0.2.14
scipy==1.11.4
flask==3.0.2